+++ compiled/src/clean-parser.d.ts
@@ -3,5 +3,7 @@ export declare class CleanParser {
     private removeEmptyTags;
     private removeEmptyTextNodes;
     private setOptions(options);
-    parse(htmlNodes: iHtmlElement[], options: iCleanOptions): void;
+    private parseAndRemoveEmptyText(index, nodes);
+    private parseAndRemoveEmptyTags(index, nodes);
+    parse(htmlNodes: iHtmlElement[], options?: iCleanOptions): iHtmlElement[];
 }