+++ compiled/src/utility.d.ts
@@ -1,5 +1,7 @@
 export declare class Utility {
     removeWhitespace(text: string): string;
+    isWhitespace(ch: string): boolean;
+    textOnlyContainsWhitespace(text: string): boolean;
     isLetter(ch: string): boolean;
     isStartOfTag(ch: string, nextCh: string): boolean;
     isEndOfTag(ch: string, nextCh: string): boolean;