+++ src/clean-parser.ts
@@ -1,4 +1,6 @@
 import { iHtmlElement, iCleanOptions } from './interfaces';
+import { ELEMENT_TYPES, TAG_TYPES } from './constants';
+import { utility } from './utility';
 
 /**
  * Clean up the array of node elements returned from the HtmlParser.parse()
@@ -29,9 +31,76 @@ export class CleanParser {
   }
 
 
-  parse(htmlNodes: iHtmlElement[], options: iCleanOptions) {
-    this.setOptions(options);
+  /**
+   * Remove text nodes that only contain whitespace
+   * 
+   * @private
+   * @param {number} index 
+   * @param {iHtmlElement[]} nodes 
+   * @returns {void} 
+   * @memberof CleanParser
+   */
+  private parseAndRemoveEmptyText(index: number, nodes: iHtmlElement[]): void {
+    if(index >= nodes.length) {
+      return;
+    }
+    let node = nodes[index];
+    if(node.type === ELEMENT_TYPES.TEXT 
+    && utility.textOnlyContainsWhitespace(node.data)) {
+      nodes.splice(index, 1);
+      index--;
+    } else if(node.type === ELEMENT_TYPES.TAG && node.children
+    && node.children.length > 0) {
+      this.parseAndRemoveEmptyText(0, node.children);
+    }
+    this.parseAndRemoveEmptyText(++index, nodes);
+  }
 
-    
+
+  /**
+   * Remove tag elements that are empty
+   * 
+   * @private
+   * @param {number} index 
+   * @param {iHtmlElement[]} nodes 
+   * @returns {void} 
+   * @memberof CleanParser
+   */
+  private parseAndRemoveEmptyTags(index: number, nodes: iHtmlElement[]): void {
+    if(index >= nodes.length) {
+      return;
+    }
+    let node = nodes[index];
+    if(node.type === ELEMENT_TYPES.TAG) {
+      if(node.children && node.children.length > 0) {
+        this.parseAndRemoveEmptyTags(0, node.children);
+      } 
+      let noChildern = (!node.children || node.children.length <= 0);
+      if(node.tagType === TAG_TYPES.DEFAULT && noChildern) {
+        nodes.splice(index, 1);
+        index--;
+      }
+    }
+    this.parseAndRemoveEmptyTags(++index, nodes);
+  }
+
+
+  /**
+   * Parse html element nodes to clean and remove unwanted tags
+   * 
+   * @param {iHtmlElement[]} htmlNodes 
+   * @param {iCleanOptions} [options] 
+   * @returns {iHtmlElement[]} 
+   * @memberof CleanParser
+   */
+  parse(htmlNodes: iHtmlElement[], options?: iCleanOptions): iHtmlElement[] {
+    this.setOptions(options);
+    if(this.removeEmptyTextNodes) {
+      this.parseAndRemoveEmptyText(0, htmlNodes);
+    }
+    if(this.removeEmptyTags) {
+      this.parseAndRemoveEmptyTags(0, htmlNodes);
+    }
+    return htmlNodes;
   }
 }